Created display_all_vehicles field and converted decimal to float field

+++ vehicles/models.py
@@ -5,8 +5,8 @@ from django.db import models
 
 
 class Vehicle(models.Model):
-    height = models.DecimalField(default=0)
-    diameter = models.DecimalField(default=0)
+    height = models.FloatField(default=0)
+    diameter = models.FloatField(default=0)
     mass = models.IntegerField(default=0)
     stages = models.IntegerField(default=0)
     payload_to_leo = models.IntegerField(default=0)
